added and adjusted delete CRUD functionality for backend

+++ back-end/db/schema.sql
@@ -8,7 +8,7 @@ CREATE TABLE flowers (
     id SERIAL PRIMARY KEY,
     name TEXT NOT NULL,
     scientific_name TEXT,
-    cost DECIMAL NOT NULL,
+    cost INT NOT NULL,
     color TEXT,
     watering TEXT,
     sunlight TEXT,
